requirements.txt
streamlit
biopython